Add formatting module
Add formatting module containing specs for nvim-lint and conform.

Configure nvim-lint and conform. nvim-lint and conform contain specs for
no filetype formatters, as these are to be merged in from lang modules.
Conform config creates an autocommand which substitutes formatexpr for
one provided by conform depending on the filetype formatter when the
a filetype it has a spec for is detected. Formatting can be done
visually or around motions with gq or against the whole file with
<Leader>cf; otherwise no automatic formatting is configured. nvim-lint
provides diagnostic linter suggestions on filewrite via another
autocommand.